Fix minor problem with x264 parameters

+++ server/mp4live/gui/encoder_dialog.cpp
@@ -62,8 +62,8 @@ on_EncoderSettingsDialog_response         (GtkWidget       *dialog,
 	break;
       case GUI_TYPE_FLOAT:
       case GUI_TYPE_FLOAT_RANGE:
-	pConfig->SetIntegerValue(*bptr->index,
-				 gtk_spin_button_get_value_as_int(GTK_SPIN_BUTTON(action)));
+	pConfig->SetFloatValue(*bptr->index,
+			       gtk_spin_button_get_value_as_int(GTK_SPIN_BUTTON(action)));
 	break;
       case GUI_TYPE_STRING: 
 	str = gtk_entry_get_text(GTK_ENTRY(action));
